+++ src/components/Form/CardOwnerForm.tsx
@@ -1,14 +1,15 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import Input from "./Input";
 import FormElement from "../common/FormElement";
 import { CardFormProps } from "./CardNumberForm";
+import useCardOwnerForm from "../../hooks/useCardOwnerForm";
 
 interface UserNameFormProps extends CardFormProps {
   userName: string[];
   setUserName: React.Dispatch<React.SetStateAction<string[]>>;
 }
 
-const UserNameForm = ({
+const CardOwnerForm = ({
   labelContent,
   inputCount,
   type,
@@ -18,33 +19,13 @@ const UserNameForm = ({
   onValidation,
   onFocus,
 }: UserNameFormProps) => {
-  const [, setAllInputValid] = useState(false);
-  const [errorMessage, setErrorMessage] = useState("");
-  const [inputValidity, setInputValidity] = useState({ 0: false });
   const [isFocused, setIsFocused] = useState(false);
+  const { errorMessage, setErrorMessage, updateInputValidity } = useCardOwnerForm(onValidation);
 
   const handleFocus = (type: string) => {
     onFocus(type);
     setIsFocused(true);
   };
-
-  // NOTE: 각 입력 필드의 유효성 검사 결과를 업데이트
-  const updateInputValidity = (index: number, isValid: boolean) => {
-    setInputValidity((prevValidities) => ({
-      ...prevValidities,
-      [index]: isValid,
-    }));
-  };
-
-  // NOTE: 모든 입력 필드가 유효한지 검사
-  useEffect(() => {
-    const allValid = Object.values(inputValidity).every((isValid) => isValid);
-    setAllInputValid(allValid);
-    setErrorMessage(allValid ? "" : "이름은 30자 이하의 영문 대문자여야 합니다.");
-
-    if (onValidation && allValid) onValidation(true);
-    if (onValidation && !allValid) onValidation(false);
-  }, [inputValidity]);
 
   const inputs = Array.from({ length: inputCount }, (_, index) => (
     <Input
@@ -71,4 +52,4 @@ const UserNameForm = ({
   );
 };
 
-export default UserNameForm;
+export default CardOwnerForm;
